feat: add registry into app context
