+++ src/context/AppContext.tsx
@@ -1,4 +1,5 @@
 import { createContext } from 'react';
+import { OperationRegistry } from '../services/chain/registry';
 
 export type AppContextType = {
   errorText: string;
@@ -7,6 +8,13 @@ export type AppContextType = {
   setIsReady: React.Dispatch<React.SetStateAction<boolean>>;
   isRequesting: boolean;
   setIsRequesting: React.Dispatch<React.SetStateAction<boolean>>;
+  registry: OperationRegistry;
+
+  getOpenAITools: () => unknown[];
+  executeOperation: (
+    operationType: string,
+    params: unknown,
+  ) => Promise<unknown>;
 };
 
 export const AppContext = createContext<AppContextType | undefined>(undefined);
